call cluster clean up before gathering information about ceph build
This commit helps to unblock cleaning up a cluster that was created
previously. Recently when manually cleaning up a cluster, an
exception is raised when trying to split a variable with none value
(rhbuild). Since rhbuild is not needed for cleanup, this change moves
the clean up code block to above the block where it access information
about the ceph build.