unittests: reorganize and use pytest fixtures
Move the custom_ceph_config() unit tests to a file and directory
structure that matches the import path. The code under test lives in
"utility/utils.py", so the unit tests will be "utility/test_utils.py".

Rename the fixture data file and move it to a "fixtures" directory. This
makes it easier to understand what this file is.

Operate on the fixture data file using pytest fixtures. This loosens the
coupling between the test data path and the unit test code, and makes it
easier to add future tests and fixture files later on.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>