[Lib] Add baremetal support
This patch enables the cephci to execute tests in baremetal env.
There are three major steps added to handle this integration, as follows:

1. Re-image the baremetal node to the desired OS version
2. Enable remote login and create root credentials
3. Perform wipefs on all disks except the root and reboot

The re-imaging part is now handled via Teuthology framework. For this,
magna006 node has been configured with Teuthology and its dependecies.
All the Teuthology related commands are executed on the baremetal nodes
via Magna006 (Teuthology infra node).
As mentioned, the 'reimage' command is used to reimage the baremetal node
to the desired os and for the same, other Teuthology commands are also used
to identify the current owner, unlock/lock the node, release the node etc.

To enable the remote login, the ssh_config file of the node is searched for the
'PermitRootLogin yes', if not present, the same will be added. Also the root
password is reset to a predefined one for easier root access within the framework.
The first time connection is made possible via the magna002 ssh key which is copied
to the Teuthology infra node. This helps in connecting to the node before remote login
is permitted or root credentials are set

Once the reimage and remote logins setup are complete, the disks has to be cleared for
any previous ceph data (all except the root). Wipefs command is used to handle this.
This step is followed by a reboot command.

A new required argument is also added along with the changes. In run.py, '--rhel-version'
argument is added which helps the user to specify which version of Rhel should be used to
reimage the node. This parameter is mandatory if the cloud type is baremetal, else it can
be ignored

Attaching the complete run result with the changes

Run 1 : http://magna002.ceph.redhat.com/cephci-jenkins/cephci-run-Baremetal-Integration/
Run 2 : http://magna002.ceph.redhat.com/cephci-jenkins/cephci-run-Baremetal-Integration-run2/

Signed-off-by: Pranav <prprakas@redhat.com>