[Lib] Add support for fetching images from config
Problem:
The disconnected installation requires the images to be
added to the private registry created. Currently there is
no way to accept these images.

Solution:
The solution proposed is to add the image details in the
cephci.yaml file and add support to fetch these images for
a given build_type (pacific/quincy etc.# )

Signed-off-by: Pranav Prakash <pranavprakash20@gmail.com>