[Lib] Add support for new provisioner
Objective
Create provisioning.py (provisioning) - which takes conf and environment as input, creates nodes and returns the cluster conf

Problem Statement
The current ceph ci framework uses run.py as the entry point and all the prerequisites are handled by run.py making it difficult to debug and maintain. The idea is to detach the provisioning of vm part from run.py and maintain it as an independent entity, which will help us in better maintainability and ease of debugging.

Signed-off-by: Pranav Prakash <pranavprakash20@gmail.com>