[Hardware Monitoring] [Baremetal] Add support for hardware monitoring
Problem:
Need to have a single destination for hardware monitoring
which can give the details of the machines available, current
status, locked status etc.

Solution:
Design and implement a standalone utility that can connect to
teuthology and get the information, push the data to a db.
This will help us in visualising the data using grafana

Signed-off-by: Pranav Prakash <prprakas@redhat.com>