[TestFix][TFA] Fix node reboot test scenario
Problem:
The cluster goes to health warn when the node reboot
scenario was tested. The scenario included rebooting
all the nodes and see if the cluster is healthy and is
not going to the bad or warn state. But after few node
restarts, we could see the cluster going to bad state

Reason:
The reason behind this behaviour is due to sudden reboots,
as the code reboots a node immediately after the previous
node is up and cluster check is done.

Solution:
Add an explicit timeout before every node reboot. This has
resolved the issue

Signed-off-by: Pranav Prakash <prprakas@redhat.com>