Improvement to long_running method
During unit testing, it has been observed that the method does not
honor the timeout when there is a continuous stream of reads.

For example, if the below command is executed
```
node.exec(cmd="while [ -f /ceph-qe-ready ]; do echo 'Writing...' ; done",
    long_running=True,
    timeout=10
)
```
then the timeout of 10 seconds is never honoured due to never
ending reads.

In this patch, the loop around data reading is removed and a
singular controlled loop is implemented.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>