exec_command hangs when there is no read/write occurring
The default timeout is not honoured when there are no read/write
occurring on the streams. This is the expected behavior in paramiko.

In this commit, exec_command uses long_running method instead of
calling paramiko.client.exec_command. This is done to provide us with
more control on the execution time.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>