Reduce the noise that is being generated with the new exec_command
This commit attempts to reduce the noise that is being generated
with the new exec_command.

Here, we remove gather cluster status for every orch operation as
cephadm has evolved. The verbose logging is disabled in cephadm
as it is logging to stderr.

Switching to pretty_print for cases it is required.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>