Decoupling from centralized log class
The current implementation has multiple limitations due to the
customization of logger. Some the limitations are

- mangled output when executing in parallel making debugging hard
- improper module heirachy being applied
- incorrect logging names.

This commit moves to use the native approach while still retaining
some of the enhancments required by the framework.

With this change, each test module has full control of the logging
mechanism. This would allow them to create additional handlers
without impacting other modules.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>