RHCEPHQE-17401: Refactor with context in parallel module.
The initial design populated the results only when it had
completed successfully. This could be a impact when qe is
refering the values during the context.

In this PR, the futures.as_completed is not used instead the
timeout is controlled via datetime methods. Also, results are
continued to be appended at the end of the method.

The iteration is modified to use `_futures` instead of _results.
This way we can capture details on the cancelled threads also.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>