Add customization support for IBM Cloud VPC
Multiple credential file would need to be maintained in the CI
if we follow the current ibm_vpc implementation. As part of this
commit, we add support for overriding the VPC profile via
custom-config

The following keys will be processed if the cloud type is ibmc

`ibmc_vpc` -> the name of the known VPC profile to be selected.
`ibmc_profile` -> the VM profile to be used for provisioning.

With this change, we can avoid storing multiple credential files
along with the need to have a seperate inventory maintained for
each VM profile.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>