Move away from distutils
`distutils` is deprecated from python 3.10. Due to this usage in
our repo, we unable to support higher python version. As part of
this commit, we are moving to use looseversion.LooseVersion instead
of distutils.LooseVersion

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>

Move to higher version of reportportal-client

There were some references to older python versions hence moving
to newer version of report-portal-client.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>